third_party/blink/web_tests/media/picture-in-picture/picture-in-picture-enabled.html

<!DOCTYPE html>
<title>Test Picture-in-Picture disabled by system</title>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../../resources/testdriver.js"></script>
<script src="../../resources/testdriver-vendor.js"></script>
<script src="resources/picture-in-picture-helpers.js"></script>
<body></body>
<script>
promise_test(async t => {
  assert_true(document.pictureInPictureEnabled);

  internals.settings.setPictureInPictureEnabled(false);
  assert_false(document.pictureInPictureEnabled);

  const video = await loadVideo(document, '../content/test.ogv');
  return promise_rejects_dom(t, 'NotSupportedError',
    requestPictureInPictureWithTrustedClick(video));
});
</script>
